y_offset option now available
This is a fix for issue #3 and allows for a y_offset option be used for
trace-based renderers. We have included a new example to show how this
can be used to plot various overlapping tracks. An alternative approach
is to use separate axes for each track - this option is also shown in
the example.